refactor(genres): extract genreColor into a shared util
