+++ src/pages/Genres.tsx
@@ -11,19 +11,7 @@ import { fetchGenreCatalog, filterGenresWithContent } from '../utils/library/gen
 import { libraryScopeForServer } from '../api/subsonicClient';
 import { peekGenreCatalogCache } from '../utils/library/genreCatalogCountsCache';
 import { resolveIndexKey } from '../utils/server/serverIndexKey';
-
-const CTP_COLORS = [
-  'var(--ctp-rosewater)', 'var(--ctp-flamingo)', 'var(--ctp-pink)', 'var(--ctp-mauve)',
-  'var(--ctp-red)', 'var(--ctp-maroon)', 'var(--ctp-peach)', 'var(--ctp-yellow)',
-  'var(--ctp-green)', 'var(--ctp-teal)', 'var(--ctp-sky)', 'var(--ctp-sapphire)',
-  'var(--ctp-blue)', 'var(--ctp-lavender)',
-];
-
-function genreColor(name: string): string {
-  let h = 0;
-  for (let i = 0; i < name.length; i++) h = (h * 31 + name.charCodeAt(i)) >>> 0;
-  return CTP_COLORS[h % CTP_COLORS.length];
-}
+import { genreColor } from '../utils/library/genreColor';
 
 const SCROLL_KEY = 'genres-scroll';
 const FONT_MIN_REM = 0.78;
